# Make BooleanPolynomial.variables() way faster